hveto.plot: moved function from hveto.triggers
to format column name for plot
